Rename pnc-update-dataset-xml directory to pncUpdateDatasetXml
